requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
pillow==10.4.0
email-validator==2.2.0
pydantic[email]==2.10.2
alembic==1.13.3
